Move generate_uuid from core.tests.utils to core.utils
